+++ src-docs/src/views/form_layouts/full_width.js
@@ -36,6 +36,7 @@ export default () => (
         fullWidth
       />
     </EuiFormRow>
+
     <EuiFormRow
       label="Often useful for text areas"
       fullWidth