+++ src/components/form/described_form_group/index.js
@@ -1,0 +1,1 @@
+export { EuiDescribedFormGroup } from './described_form_group';